updating a cast so it can compile on linux

+++ get_next_line.c
@@ -26,13 +26,13 @@ static char		*stresize(char **s, int in, int size)
 
 static int		fill_line(char **line, char **s, int red)
 {
-	int		len;
+	int			len;
 
 	if (!*s)
 		return (-1);
 	if (!red && !ft_strlen(*s))
 		return (0);
-	len = ft_strchr(*s, '\n') ? ft_strchr(*s, '\n') - *s : ft_strlen(*s);
+	len = ft_strchr(*s, '\n') ? ft_strchr(*s, '\n') - *s : (int)ft_strlen(*s);
 	*line = ft_strnew(len);
 	if (!*line || red == -1)
 		return (-1);
